WIP: test adjoint for different paddings
