+++ odl/util/numerics.py
@@ -27,7 +27,8 @@ import numpy as np
 from odl.util.normalize import normalized_scalar_param_list, safe_int_conv
 
 
-__all__ = ('apply_on_boundary', 'fast_1d_tensor_mult', 'resize_array')
+__all__ = ('apply_on_boundary', 'fast_1d_tensor_mult', 'resize_array',
+           'matrix_mul_axis')
 
 
 _SUPPORTED_RESIZE_PAD_MODES = ('constant', 'symmetric', 'periodic',
@@ -182,6 +183,68 @@ def apply_on_boundary(array, func, only_once=True, which_boundaries=None,
         slices[ax] = slice(start, end)
 
     return out
+
+
+def matrix_mul_axis(matrix, array, axis):
+    """Return the matrix-vector product along an axis in ``array``.
+
+    This function computes the usual matrix-vector product of ``matrix``
+    with ``array``, where the summation is performed along the
+    given ``axis``.
+
+    Parameters
+    ----------
+    matrix : array-like
+        Array with shape ``(m, n)``, acting as the matrix in the
+        product.
+    array : array-like
+        Array with ``array.ndim >= 1``, acting as right operand in
+        the product. It must fulfill ``array.shape[axis] == n``.
+    axis : int
+        Axis along which the summation in the matrix-vector product is
+        performed. It must fulfill ``-array.ndim - 1 <= axis < array.ndim``,
+        where negative values are interpreted in the usual "backwards
+        indexing" sense.
+
+    Returns
+    -------
+    prod : `numpy.ndarray`
+        Matrix-vector product along the given axis. Its shape is
+        ``(..., m, ...)``, where the value ``m`` appears at the index
+        ``axis`` and the other values are the same as in ``array.shape``.
+    """
+    matrix = np.asarray(matrix)
+    array = np.asarray(array)
+    axis, axis_in = int(axis), axis
+
+    if array.ndim == 0:
+        raise ValueError('`array` must have at least 1 dimension')
+
+    if matrix.ndim != 2:
+        raise ValueError('`matrix` must have 2 dimensions, got ndim == {}'
+                         ''.format(matrix.ndim))
+
+    if axis != axis_in:
+        raise ValueError('`axis` {} is not integer'.format(axis_in))
+
+    if not -array.ndim <= axis < array.ndim:
+        raise ValueError('`axis` {} not in the valid range {} -> {}'
+                         ''.format(axis, -array.ndim, array.ndim - 1))
+
+    if array.shape[axis] != matrix.shape[1]:
+        raise ValueError('`array.shape[axis]` == {} and `matrix.shape[1] '
+                         '== {} do not match'.format(array.shape[axis],
+                                                     matrix.shape[1]))
+
+    out_arr = np.tensordot(matrix, array, axes=[[1], [axis]])
+    try:
+        # np.moveaxis was added in Numpy 1.11
+        return np.moveaxis(out_arr, 0, axis)
+    except AttributeError:
+        # Fallback for older Numpy versions
+        newshape = list(range(1, array.ndim))
+        newshape.insert(axis, 0)
+        return np.transpose(out_arr, newshape)
 
 
 def fast_1d_tensor_mult(ndarr, onedim_arrs, axes=None, out=None):
